Cache property descriptors
`Reflect.getOwnPropertyDescriptor` is moderately expensive and it is
invoked on every access of a non-primitive attribute. Since they don’t
change often, it makes sense to cache them. The cache is a two-level
map of target -> property name -> property descriptor. The first mapping
is a WeakMap so it will not interfere with garbage collection.