Don’t fire callback when set doesn’t change prop
On `set` do a strict equality test between the previous and new value,
and only fire the change callback if the value actually changed. As
part of this fix, we also “unwrap” proxied values before assigning them
to the target. This avoids “polluting” the underlying, wrapped object
with proxies (which applications may access separately).