Tweak/Add benchmarks
Reorganize the benchmarks into a `native` and `on-change` suite for
easier comparison. In addition, removed the loop in the `save` method.
Since the implementation of the callback is outside the control of this
utility, imo it makes sense to compare performance with a “no-op”
callback. Proxies add a fair amount of overhead compared with native
access, but the 1000-iteration loop was more or less overwhelming this
overhead…